Fix extra unresponsive messages being sent when SMS response was received from a different contact
When a user replied from contact A, we would send the 'unresponsive' SMS
to contact B along with their next rollcall. We need to search for
responses to a rollcall, not a specific contact.

Also make there join to replies not matching more explicit by using whereNull on reply ID